+++ CHANGELOG.md
@@ -361,3 +361,15 @@
 - K_asset file generation command added
 
 
+## 1.5.5
+
+- automatic route added with repo based module generation
+- pubspec edit and config all neccessary packages and other stuff
+
+
+## 1.5.6
+- update changelog file
+
+## 1.5.7
+- auto route generation issue resolve
+